angrylion HLE: apply G_TEXTURE scale in fixed point, not as a float multiply
The per-vertex texture coordinate was formed as a float multiply of the raw
S10.5 texel coordinate by a float-converted G_TEXTURE scale (scale / 65536.0f
in the G_TEXTURE handler, then st * scale as float). The RSP applies the
S0.16 G_TEXTURE scale to the texel coordinate in fixed point (vmudn/vmadm
against the scale vector), keeping an S10.5 result.

Carry the G_TEXTURE scale as the raw u16 S0.16 value and compute the texture
coordinate as (st * scale) >> 16 in 64-bit fixed point, matching the RSP, with
no float scale factor and no round-trip through a normalised float.

HLE encoder/frontend only; the rasterizer core is unchanged.

+++ mupen64plus-video-angrylion/rdp_emit_frontend.h
@@ -43,7 +43,7 @@ typedef struct GSPState
     int   combined_valid;
 
     BridgeViewport viewport;
-    float          tex_scale_s, tex_scale_t;
+    unsigned int   tex_scale_s, tex_scale_t; /* raw S0.16 from G_TEXTURE */
     int            tex_tile, tex_level, tex_w, tex_h;
 
     unsigned int   geometry_mode;
@@ -69,7 +69,7 @@ void gsp_combine_matrices(GSPState *s);
 
 /* viewport (addr -> N64 Vp struct in RDRAM), and texture scale state */
 void gsp_set_viewport(GSPState *s, const unsigned char *rdram, unsigned int addr);
-void gsp_set_texture(GSPState *s, float scale_s, float scale_t,
+void gsp_set_texture(GSPState *s, unsigned int scale_s, unsigned int scale_t,
                      int tile, int level, int tex_w, int tex_h);
 
 /* load+transform n vertices starting at index v0 from RDRAM addr */
